backend/_home/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
algoliasearch==4.16.0
algoliasearch-django==4.0.0
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
multidict==6.2.0
mysqlclient==2.2.7
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.3.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.18.3
